fix(matching): corrige cast silencieux et ajoute test miroir pour PENDING_BID
- Remplace (int) upcoming par Math.toIntExact pour lever ArithmeticException
  si la valeur dépasse Integer.MAX_VALUE, cohérent avec WalletDeletionImpactContributor
- Ajoute pendingBid_mergesBothRoles : prouve que bothRoles() agrège expéditeur
  et voyageur pour BidStatus.PENDING, symétrique à parcelInTransit_mergesBothRoles

+++ src/main/java/com/yadony/api/matching/MatchingDeletionImpactContributor.java
@@ -51,7 +51,7 @@ public class MatchingDeletionImpactContributor implements UserDeletionImpactCont
         long upcoming = announcementRepository.countByTravelerIdAndStatusIn(userId, UPCOMING);
         if (upcoming > 0) {
             findings.add(ImpactFinding.plain(
-                    ImpactSeverity.WARNING, "UPCOMING_ANNOUNCEMENT", (int) upcoming));
+                    ImpactSeverity.WARNING, "UPCOMING_ANNOUNCEMENT", Math.toIntExact(upcoming)));
         }
 
         return findings;
